refactor: Move k8s resource readers from validate package
Move the resource reader helpers from `pkg/validate/core.go` to
new `pkg/core` package and export them. These functions are
generic utilities for reading Kubernetes resources and are not
specific to the validate package.
